# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agrisiti Academy API",
        "description": "REST and WebSocket API for the Agrisiti Academy e-learning platform. Covers student registration, course management, progress tracking, tests, assignments, notes, messaging, community forum, notifications, and the Agri chatbot (AI-powered onboarding assistant with real-time Reverb WebSocket streaming, SSE fallback, and course recommendations). See the Chatbot tag for the complete user journey, quiz questions, and 3-tier transport documentation.",
        "contact": {
            "email": "admin@agrisiti.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://academy-backends.agrisiti.com",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/career/apprenticeships/{slot}/express-interest": {
            "post": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Express interest in an apprenticeship slot; the learner's certificate for the required course (if any) is attached automatically",
                "operationId": "c3db69d18009c10c3bdf54a8ebb1c0fd",
                "parameters": [
                    {
                        "name": "slot",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Interest recorded"
                    },
                    "400": {
                        "description": "Already applied, slot closed, or required course not completed"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/career/my-applications": {
            "get": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "List the authenticated learner's apprenticeship applications and their status",
                "operationId": "8aabe30b348178cf77a2c99b9ce55f4a",
                "responses": {
                    "200": {
                        "description": "Learner's applications"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/org/slots/{slot}/applicants": {
            "get": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "List applicants for a slot the organisation owns, each with their attached certificate",
                "operationId": "076d8870c7416dce988e155f6d817895",
                "parameters": [
                    {
                        "name": "slot",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Applicants list"
                    },
                    "403": {
                        "description": "Not the owning organisation"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/org/applications/{apprenticeship}/review": {
            "post": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Accept or reject a learner's application; the learner is notified of the outcome",
                "operationId": "2510fccc083945fe6209afa9f10b3d3f",
                "parameters": [
                    {
                        "name": "apprenticeship",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "accepted",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application reviewed"
                    },
                    "403": {
                        "description": "Not the owning organisation"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/apprenticeships/{apprenticeship}/logs": {
            "get": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Full log history for an apprenticeship as a timeline, with days since acceptance that have no log flagged",
                "operationId": "758c1f098007e517ee01e412dcd876c5",
                "parameters": [
                    {
                        "name": "apprenticeship",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timeline: logged days plus 'no log submitted' placeholders for missed days",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "has_log": {
                                                        "type": "boolean"
                                                    },
                                                    "attended": {
                                                        "type": "boolean",
                                                        "nullable": true
                                                    },
                                                    "activity_description": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not the apprentice, host organisation, or an admin"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Career Pathways"
                ],
                "summary": "Log today's attendance/activity for an active apprenticeship (learner or host organisation)",
                "operationId": "0ab402619a9a9602546f2b798cc0a08b",
                "parameters": [
                    {
                        "name": "apprenticeship",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "log_date": {
                                        "description": "Defaults to today",
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "attended": {
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "activity_description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Log recorded (or updated, if one already existed for that date)"
                    },
                    "403": {
                        "description": "Not the apprentice, host organisation, or an admin"
                    },
                    "400": {
                        "description": "Apprenticeship is not in an active (accepted) state"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token from /api/login"
            }
        }
    },
...
}